+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fix formatting date time property with UTC time zone
+
 ## [2.9.0] - 2022-12-23
 
 ### Changed